% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{AddHistory}
\alias{AddHistory}
\alias{addhistory}
\title{Amend the history of a proxy object}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
AddHistory(x, newhist, date = TRUE)
}
\arguments{
\item{x}{a \code{"pTs"} or \code{"pField"} object.}

\item{newhist}{character string with the history information to be added.}

\item{date}{logical, whether or not to append the current date to the
history attribute; defaults to \code{TRUE}.}
}
\value{
the \code{"pTs"} or \code{"pField"} object with the amended
history.
}
\description{
This function amends the history information of a \code{"pTs"} or
\code{"pField"} object by adding a given character string (e.g., a comment)
to the \code{"history"} attribute of the object together with the date
information of this action.
}
\examples{
# Create empty pTs object:
x <- pTs(lat = -75, lon = 0, time = 1,
         history = sprintf("Object created."))

# Check history
GetHistory(x)

# Some changes made
x <- AddHistory(x, newhist = "Changed xyz.")

# Check new history
GetHistory(x)
}
\seealso{
Other history: \code{\link{GetHistory}}
}
\author{
Thomas Laepple
}
\concept{history}
