% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-fields.R
\name{ApplyFields}
\alias{ApplyFields}
\title{Apply function on two fields}
\usage{
ApplyFields(fld1, fld2, FUN, newtime = NULL, ...)
}
\arguments{
\item{fld1}{a \code{"pField"} object.}

\item{fld2}{a \code{"pField"} object.}

\item{FUN}{the function to be applied on \code{fld1} and \code{fld2}.}

\item{newtime}{Specify the observation time point corresponding to the
resulting \code{"pField"} object. For \code{NULL} (the default), the average
of the time points in \code{fld1} is used.}

\item{...}{further arguments passed on to \code{FUN}.}
}
\value{
a \code{"pField"} object with the results of the function applied.
}
\description{
Apply a function, which expects two arguments, row-wise, i.e. across space,
on two \code{"pField"} objects. The result of this gives a \code{"pField"}
object with only one time step.
}
\details{
Note that the currect implementation expects both fields to have the exact
same time basis, latitudes and longitudes; any differences will result in an
error.
}
\examples{
x <- pField(data = array(rnorm(10 * 10 * 100), dim = c(10, 10, 100)),
            time = 1 : 100, lat = 1 : 10, lon = 1 : 10)
y <- pField(data = array(rnorm(10 * 10 * 100), dim = c(10, 10, 100)),
            time = 1 : 100, lat = 1 : 10, lon = 1 : 10)
z <- ApplyFields(x, y, cor)
}
\author{
Thomas Münch
}
