% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-fields.R
\name{ApplyTime}
\alias{ApplyTime}
\alias{applytime}
\title{Apply a function across time}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
ApplyTime(data, FUN, newtime = NULL, ...)
}
\arguments{
\item{data}{a \code{"pField"} or \code{"pTs"} object.}

\item{FUN}{the function to be applied.}

\item{newtime}{the observation time point(s) of the result. For \code{NULL}
(the default), the average of the time points in \code{data} is used,
assuming the applied function yields an aggregated result. If the result of
\code{FUN} yields more than one time step, you need to provide the
respective new time axis here.}

\item{...}{further arguments passed on to \code{FUN}.}
}
\value{
a \code{"pField"} or \code{"pTs"} object with the results of the
function applied.
}
\description{
Apply a function across the temporal dimension of a \code{"pField"} or
\code{"pTs"} object by using \code{apply} on the columns of the
object. Depending in the output of the applied function, this gives a
\code{"pField"} or \code{"pTs"} object with only one, or with several time
steps.
}
\examples{
x <- pField(data = array(rnorm(10 * 10 * 100), dim = c(10, 10, 100)),
            time = 1 : 100, lat = 1 : 10, lon = 1 : 10)
y <- ApplyTime(x, mean)

# subset incompletely to create a pTs object
x.pts <- x[, 1 : 15]
y.pts <- ApplyTime(x.pts, mean) # is the same as y[, 1 : 15]
}
\author{
Thomas Laepple
}
