% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{Decor.pField}
\alias{Decor.pField}
\title{Decorrelation of field}
\usage{
Decor.pField(
  field,
  lat1,
  lon1,
  lat2,
  lon2,
  return.scatter = FALSE,
  verbose = FALSE,
  print.progress = NULL,
  ...
)
}
\arguments{
\item{field}{a \code{"pField"} object.}

\item{lat1}{numeric vector of length 1 with the latitude of the target
location for which the decorrelation length is to be calculated. Omit to
analyse the entire field.}

\item{lon1}{numeric vector of length 1 with the longitude of the target
location for which the decorrelation length is to be calculated. Omit to
analyse the entire field.}

\item{lat2}{optional numeric vector of length 1 with a second latitude to
define a coordinate region for which decorrelation lengths are to be
correlated. OBS! Method to select a region from the \code{"pField"} object
is not yet implemented!}

\item{lon2}{Same as \code{lat2} for a second longitude value.}

\item{return.scatter}{logical; if \code{TRUE}, a data frame with distance-
correlation pairs is returned for each target location additionally to the
decorrelation length. Defaults to \code{FALSE}.}

\item{verbose}{logical; if \code{TRUE}, informative messages are printed on
the target location selection and the decorrelation length
estimation. Defaults to \code{FALSE}.}

\item{print.progress}{numeric value to specify an index interval at which
messages on the computation progress are printed (useful for large analysed
fields). Ignored for \code{NULL} (the default).}

\item{...}{further arguments passed on to the correlation function
\code{\link{cor.pTs}}.}
}
\value{
Decorrelation lengths are returned in units of km. The return
structure depends on the supplied input:
\itemize{
\item For a single target point (lat1-lon1 pair specified), the returned
decorrelation length is a numeric length-1 vector.
\item For analysing a region (via specifying lat2 and lon2) or the entire
field (no lat-lon's specified), the returned decorrelation length is an
object of class \code{"pField"} of the same coordinate structure as the
analysed region/field.
\item For \code{return.scatter = TRUE}, the return object is a 2-element
list where the first list element is the decorrelation length result as
defined above, and the second list element is a list of data frames for each
analysed location where each data frame has elements \code{d} and \code{cor}
with all distance-correlation value pairs for this location.
}
}
\description{
Calculate the decorrelation length of a \code{"pField"} data set for given
target locations.
}
\details{
The decorrelation length estimation is based on fitting an exponential of
the form \code{exp(-x/tau)}, where \code{x} is distance and \code{tau} the
fit parameter, i.e. the decorrelation length, to the distance-correlation
relationship as obtained from correlating the time series at the target
location with the time series from all other grid points of the input
\code{"pField"}. The fit is performed using the built-in R function
\code{\link[stats]{nlm}}.
}
\seealso{
\code{\link{cor.pTs}}
}
\author{
Thomas Münch
}
