% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert between degree and radian}
\source{
This is a copy of the function of the same name in package
\code{"ecustools"}; see
\url{https://github.com/EarthSystemDiagnostics/ecustools}.
}
\usage{
deg2rad(x, inverse = FALSE)
}
\arguments{
\item{x}{a numeric vector of coordinate values to be converted.}

\item{inverse}{logical; convert from degree to radian (the default) or from
radian to degree (\code{inverse = TRUE}).}
}
\value{
numeric vector of the same length as \code{x} with the converted
coordinate values.
}
\description{
Convert coordinates from degree to radian and vice versa.
}
\examples{
deg2rad(deg2rad(-75), inverse = TRUE)
}
\author{
Thomas Münch
}
