% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-fields.R
\name{FirstElement}
\alias{FirstElement}
\title{first element of a vector}
\usage{
FirstElement(x)
}
\arguments{
\item{x}{vector}
}
\value{
first element of X
}
\description{
first element of a vector
}
\author{
Thomas Laepple
}
