% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-fields.R
\name{GetDistance}
\alias{GetDistance}
\title{Find distance between points on Earth}
\source{
This is a copy of the function of the same name in package
\code{"ecustools"}; see
\url{https://github.com/EarthSystemDiagnostics/ecustools}.
}
\usage{
GetDistance(lat0, lon0, lat, lon, get.nearest = FALSE, verbose = FALSE)
}
\arguments{
\item{lat0}{numeric vector of length 1 with the latitude (in degree) of the
target location relative to which distances are to be calculated.}

\item{lon0}{numeric vector of length 1 with the longitude (in degree) of the
target location.}

\item{lat}{numeric vector of latitudes (in degree) of the positions for
which distances from the target position are to be calculated.}

\item{lon}{numeric vector of longitudes (in degree) of the positions for
which distances from the target position are to be calculated. Must have the
same length as \code{lat}.}

\item{get.nearest}{logical; specify whether the given 'lat0' - 'lon0' pair
shall be used _as is_ as the target location (\code{FALSE}), or instead the
nearest point of the positions specified by 'lat' and 'lon' (\code{TRUE}).
Selection of the nearest point is currently implemented with
\code{\link{MinimizeSpherical}}.}

\item{verbose}{logical; if \code{TRUE} an informative message regarding the
selected target location is printed.}
}
\value{
Numeric vector of the same length as \code{lat} with the distances
in km from the target location.
}
\description{
This is a general function to calculate the distance on Earth between a
target location and all points in a field of coordinates. It provides a
straighforward interface to the function \code{\link[geosphere]{distGeo}}
from package \code{geosphere}.
}
\details{
This is a wrapper function whose purpose it is to provide a straighforward
interface to the function \code{\link[geosphere]{distGeo}}; this includes:
\itemize{
\item direct input of latitude/longitude values which are then shaped into
      the appropriate structure for \code{\link[geosphere]{distGeo}};
\item longitude values are automatically confined to [-180, 180] as needed by
      \code{\link[geosphere]{distGeo}};
\item if required, one can select the point closest to the target location as
      the reference for the distance calculations.}
}
\examples{
# some coordinates
lat0 <- -75
lon0 <- 0
lat <- seq(-74, -80, -2)
lon <- seq(-2, 7, 3)

d <- GetDistance(lat0, lon0, lat, lon, verbose = TRUE)
range(d)

d <- GetDistance(lat0, lon0, lat, lon, get.nearest = TRUE, verbose = TRUE)
range(d)
}
\seealso{
\code{\link{GetDistanceField}}
}
\author{
Thomas Münch
}
