% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-fields.R
\name{GetDistanceField}
\alias{GetDistanceField}
\title{Calculate distance between grid points}
\usage{
GetDistanceField(
  field,
  latlon.field = NULL,
  lat,
  lon,
  get.nearest = FALSE,
  index = NULL,
  verbose = FALSE
)
}
\arguments{
\item{field}{a \code{"pField"} object.}

\item{latlon.field}{the optional result of calling
\code{\link{GetLatLonField}} for \code{field}; see details. For \code{NULL}
(the default), \code{\link{GetLatLonField}} is called internally.}

\item{lat}{numeric vector of length 1 with the latitude of the target
location relative to which distances are to be calculated.}

\item{lon}{numeric vector of length 1 with the longitude of the target
location.}

\item{get.nearest}{logical to specify whether the given latitude - longitude
pair shall be used *as is* as the target location (\code{FALSE}), or instead
the nearest grid point in \code{field} (\code{TRUE}).}

\item{index}{an integer array index of the input \code{"pField"} for
specifying the target location. Setting this will override any \code{lat},
\code{lon} settings, while the default \code{NULL} signals to use the
specified \code{lat}, \code{lon} values.}

\item{verbose}{logical; if \code{TRUE} an informative message regarding the
selected target location is printed.}
}
\value{
Numeric vector with distances in km of the grid points of
\code{"field"} relative to the target location following the spatial
ordering of a \code{"pField"} object.
}
\description{
Calculate the distances between a specified target location and all grid
points of a given \code{"pField"}. The calculation is based on the function
\code{\link[geosphere]{distGeo}}.
}
\details{
\code{latlon.field} can be supplied for efficiency if
\code{GetDistanceField} needs to be called many times, e.g., to calculate
the distances from several target locations for the same coordinate field.
}
\examples{
field <- pField(lat = seq(0, 90, 10), lon = c(-10, 0, 10))

d <- GetDistanceField(field, lat = 5, lon = 5, verbose = TRUE)
range(d)

d <- GetDistanceField(field, lat = 5, lon = 5,
                      get.nearest = TRUE, verbose = TRUE)
range(d)

d <- GetDistanceField(field, index = 1, verbose = TRUE)
range(d)
}
\seealso{
\code{\link{GetLatLonField}}
}
\author{
Thomas Münch
}
