% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{GetHistory}
\alias{GetHistory}
\alias{gethistory}
\title{Get history of a proxy object}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
GetHistory(data)
}
\arguments{
\item{data}{a \code{"pTs"} or \code{"pField"} object.}
}
\value{
Character string with the \code{"history"} attribute of \code{data}.
}
\description{
The function returns the \code{"history"} attribute of a \code{"pTs"} or
\code{"pField"} object.
}
\examples{
# Create empty pTs object:
x <- pTs(lat = -75, lon = 0, time = 1,
         history = sprintf("Object created."))

GetHistory(x)
}
\seealso{
Other history: 
\code{\link{AddHistory}()}
}
\author{
Thomas Laepple
}
\concept{history}
