% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{GetLat}
\alias{GetLat}
\alias{getlat}
\title{Get latitudes of proxy object}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
GetLat(data)
}
\arguments{
\item{data}{a \code{"pTs"} or \code{"pField"} object.}
}
\value{
a numeric vector of the latitudes of the \code{"pTs"} or
\code{"pField"}  object.
}
\description{
The function returns the \code{"lat"} attribute of a \code{"pTs"} or
\code{"pField"} object, i.e. the latitudes of the proxy sampling locations,
as a numeric vector.
}
\author{
Thomas Laepple
}
