% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{GetLon}
\alias{GetLon}
\alias{getlon}
\title{Get longitudes of proxy object}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
GetLon(data)
}
\arguments{
\item{data}{a \code{"pTs"} or \code{"pField"} object.}
}
\value{
a numeric vector of the longitudes of the \code{"pTs"} or
\code{"pField"}  object.
}
\description{
The function returns the \code{"lon"} attribute of a \code{"pTs"} or
\code{"pField"} object, i.e. the longitudes of the proxy sampling locations,
as a numeric vector.
}
\author{
Thomas Laepple
}
