% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{GetName}
\alias{GetName}
\alias{getname}
\title{Get name of proxy object}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
GetName(data)
}
\arguments{
\item{data}{a \code{"pTs"} or \code{"pField"} object.}
}
\value{
Character string with the \code{"name"} attribute of \code{data}.
}
\description{
The function returns the \code{"name"} attribute of a \code{"pTs"} or
\code{"pField"} object.
}
\examples{
getname(c(a = 2))
GetName(c(a = 2))
}
\author{
Thomas Laepple
}
