% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{is.pField}
\alias{is.pField}
\title{pField object}
\usage{
is.pField(object)
}
\arguments{
\item{object}{object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{object} is a
\code{"pField"} or not.
}
\description{
Check if object is a \code{"pField"} object.
}
\examples{
# Some pField object
x <- pField(lat = seq(-75, -80, -5), lon = c(0, 135, 215), time = 1 : 10)

is.pField(x)
is.pField(1 : 10)
}
\author{
Thomas Laepple
}
