% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{is.pTs}
\alias{is.pTs}
\title{pTs object}
\usage{
is.pTs(object)
}
\arguments{
\item{object}{object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{object} is a
pTs or not.
}
\description{
Test for pTs object
}
\author{
Thomas Laepple
}
