% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-fields.R
\name{MinimizeSpherical}
\alias{MinimizeSpherical}
\title{Minimize difference between coordinates}
\usage{
MinimizeSpherical(lat0, lon0, lat, lon, return.coordinates = FALSE)
}
\arguments{
\item{lat0}{a length-1 vector with the latitude [degree] of the target
position.}

\item{lon0}{a length-1 vector with the longitude [degree] of the target
position.}

\item{lat}{numeric vector of latitudes [degree] from a field of coordinates
for which the point closest to the target is to be determined.}

\item{lon}{numeric vector of longitudes [degree] from a field of coordinates
for which the point closest to the target is to be determined. Must have the
same length as \code{lat}.}

\item{return.coordinates}{logical; shall the actual coordinates of the
position closest to the target be returned? Defaults to \code{FALSE}, which
returns just the vector index of the closest position.}
}
\value{
per default, the index position in \code{lat} (or \code{lon}) of the
position closest to the target; or the actual coordinates of this position
as a length-2 numeric vector (for \code{return.coordinates = TRUE}).
}
\description{
This function finds the position in a latitude/longitude field that is
closest to a given target position by minimizing the root mean square
deviation of their position vectors in spherical coordinates.
}
\examples{
# some coordinates
lat0 <- -75
lon0 <- 0
lat <- seq(-74, -80, -2)
lon <- seq(-2, 7, 3)

# get closest point to lat0/lon0 in degree
MinimizeSpherical(lat0, lon0, lat, lon,
                  return.coordinates = TRUE)
}
\author{
Thomas Münch
}
