% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-methods.R
\name{Ops.pTs}
\alias{Ops.pTs}
\title{pTs operator method}
\source{
Function adapted from "proxytype.R" in paleolibary/src/
}
\usage{
\method{Ops}{pTs}(e1, e2)
}
\arguments{
\item{e1}{an object of class \code{"pTs"}.}

\item{e2}{an object, optionally of class \code{"pTs"} (or missing for
unary operator).}
}
\value{
an object of class \code{"pTs"} with the result of the operator
applied on \code{e1} and \code{e2}.
}
\description{
Group generic method for the \code{Ops} group of functions to be applied on
\code{"pTs"} objects.
}
\seealso{
\code{?Ops}
}
\author{
Thomas Laepple; adapated by Thomas Münch
}
