% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pField.R
\name{pField}
\alias{pField}
\title{Create a pField object}
\source{
Function copied from "proxytype.R" in paleolibary/src/
}
\usage{
pField(data = NULL, time = 1, lat = NA, lon = NA, name = "",
  history = "", date = TRUE, kTol = 1/400)
}
\arguments{
\item{data}{input data, usually a three-dimensional gridded data array. If
\code{NULL}, or a single value, an empty data array, or a data array filled
with the single value, is created with dimensions according to the
information (i.e., lengths) in \code{time}, \code{lat} and \code{lon}.}

\item{time}{vector of observation time points, must be equidistant.}

\item{lat}{vector of latitudes.}

\item{lon}{vector of longitudes.}

\item{name}{character string with the name of the \code{"pField"} object to
be generated.}

\item{history}{character string to append to the history attribute of the
\code{"pField"} object (optional).}

\item{date}{logical, whether or not to append the current date to the
history attribute; defaults to \code{TRUE}.}

\item{kTol}{tolerance to check for equidistance of time steps; defaults to
1/400.}
}
\value{
A two-dimensional array of class \code{"pField"}.
}
\description{
This function takes supplied vectors of time points, latitudes and
longitudes to convert, or create, a three-dimensional gridded data array
into a \code{"pField"} object.
}
\details{
For a supplied gridded data array, the function shapes the data array into a
two-dimensional array with columns corresponding to the spatial dimension
(latitude-longitude pair) and rows corresponding to the time points of the
observations, i.e. similar to the structure of netcdf files. For this, the
total number of observation points in \code{data} must match the product of
the lengths of \code{time}, \code{lat} and \code{lon}. Each time series at
a spatial location is converted to a time series \code{\link[stats]{ts}}
object; and longitudes and latitudes are added to the new array as
attributes. If no input array is supplied, an empty three-dimensional
gridded data array is created from the information provided in \code{time},
\code{lat} and \code{lon}, and converted in a similar manner to a 2D
\code{"pField"} object.
}
\author{
Thomas Laepple; adapated by Thomas Münch
}
