% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{pField2df}
\alias{pField2df}
\title{Convert field to data frame}
\usage{
pField2df(data, lat.min = NULL, lat.max = NULL, lon.min = NULL,
  lon.max = NULL)
}
\arguments{
\item{data}{a \code{"pField"} object with one timestep.}

\item{lat.min}{set the minimum latitude for the output; per default the
minimum in \code{data} is used.}

\item{lat.max}{set the maximum latitude for the output; per default the
maximum in \code{data} is used.}

\item{lon.min}{set the minimum longitude for the output; per default the
minimum in \code{data} is used.}

\item{lon.max}{set the maximum longitude for the output; per default the
maximum in \code{data} is used.}
}
\value{
A data frame with the three columns \code{lat} (latitudes),
  \code{lon} (longitudes) and \code{dat} (the values of \code{data} at the
  coordinate positions).
}
\description{
Convert a static \code{"pField"} object, i.e. including only one time step,
to a data frame with data, latitude and longitude columns, with the
possibility to cut out a specified latitude-longitude region.
}
\author{
Thomas Münch
}
