% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTs.R
\name{pTs}
\alias{pTs}
\title{Create a pTs object}
\source{
Function copied from "proxytype.R" in paleolibary/src/.
}
\usage{
pTs(data, time = NULL, lat = NA, lon = NA, name = "",
  history = "", date = TRUE, kTol = 1/400)
}
\arguments{
\item{data}{a vector, matrix or ts object to be converted. If missing, a
matrix, filled with \code{NA} values, with dimensions according to the
lengths of \code{time} and \code{name} is created. For this, \code{time}
needs to be non-\code{NULL}.}

\item{time}{vector of time points corresponding to the observations in the
proxy time series; must be equidistant. If \code{NULL} (the default), a
vector of length \code{length(data)} (or \code{nrow(data)}) is created
assuming equidistant time steps with stepsize 1. For this, \code{data}
needs to be non-\code{NULL}.}

\item{lat}{numeric vector giving the latitude(s) of the sampling location(s)
of the proxy time series.}

\item{lon}{vector of length 1 giving the longitude(s) of the sampling
location(s) of the proxy time series.}

\item{name}{vector of character strings with the names of the \code{"pTs"}
object to be generated (optional).}

\item{history}{character string to append to the history attribute of the
\code{"pTs"} object (optional).}

\item{date}{logical, whether or not to append the current date to the
history attribute; defaults to \code{TRUE}.}

\item{kTol}{tolerance to check for equidistance of time steps; defaults to
1/400.}
}
\value{
An object of class \code{"pTs"}.
}
\description{
This function takes a base R vector, matrix, or ts object and converts it to
a \code{"pTs"} object.
}
\details{
\code{"pTs"} (\code{"proxy time series"}) objects are enhanced time-series
\code{\link[stats]{ts}} objects. \code{pTs()} adds attributes such as
longitude and latitude to a time series vector/time series vectors and
assigns the class \code{"pTs"} to the resulting object. More than one time
series vectors need to have the same time basis and can either stem from the
same location (i.e. different data sets/proxies recorded at the same site,
thus \code{lat}, \code{lon} are of length 1), or it can be the same proxy
data recorded at different sites (i.e. \code{lat}, \code{lon} are vectors of
the same length > 1). The latter can be used for irregular data fields which
cannot be shaped into a \code{\link{pField}} object.
}
\examples{
# 10 data sets from the same site (e.g. Kohnen Station, Antarctica)

dat <- matrix(rnorm(10 * 100), nrow = 100, ncol = 10)
dat.pts <- pTs(dat, lat = -75, lon = 0,
               name = "10 proxy time series from Kohnen.")

# time vector is automatically created
time(dat.pts)

# supply a specific time vector
dat.pts <- pTs(dat, time = 1901 : 2000, lat = -75, lon = 0,
               name = "10 proxy time series from Kohnen.")
time(dat.pts)

# supply data set names
dat.pts <- pTs(dat, time = 1901 : 2000, lat = -75, lon = 0,
               name = paste("proxy", 1 : 10))

# 10 data sets from different sites
# (= irregular grid which cannot be shaped into a pField)

lat <- seq(-75, -85, length.out = 10)
lon <- seq(0, 90, length.out = 10)

dat.pts <- pTs(dat, time = 1901 : 2000, lat = lat, lon = lon,
               name = "Proxy time series from 10 sites.")
}
\seealso{
\code{\link{pField}}
}
\author{
Thomas Laepple; adapated by Thomas Münch
}
