% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-fields.R
\name{SelSpace3D}
\alias{SelSpace3D}
\alias{selspace.3D}
\title{SelSpace3D}
\usage{
SelSpace3D(
  data,
  lat1,
  lon1,
  SBOX = 5,
  tolLon = 10,
  bNN = FALSE,
  timeindexNA = 1
)
}
\arguments{
\item{data}{pField object}

\item{lat1}{vector length 1}

\item{lon1}{vector legnth 1}

\item{SBOX}{vector length 1, defaults to 5}

\item{tolLon}{vector length 1, defaults to 10}

\item{bNN}{logical, defaults to FALSE}

\item{timeindexNA}{at which timestep should the Next Neighbour algorithm
search for missing values}
}
\value{
\code{\link{pTs}} object
}
\description{
Function to interpolate a field to a given point specified by
  latitude and longitude. It uses the nearest neighbour if the adjancents
  points are missing, if not it uses bilinear interpolation
}
\author{
Thomas Laepple
}
