% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{str.pField}
\alias{str.pField}
\title{Compactly display the structure of a pField object}
\usage{
\method{str}{pField}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"pField"}.}

\item{...}{further arguments passed on to \code{\link[utils]{str}}.}
}
\description{
Method function to compactly display the internal structure of a
\code{"pField"} object, a diagnostic function and an alternative to
\code{summary}; performed with the \code{str} method generic to the
\code{"oclass"} of \code{object}.
}
\seealso{
\code{\link[utils]{str}}
}
\author{
Thomas Münch
}
