% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing-methods.R
\name{[.pField}
\alias{[.pField}
\title{Extract or replace parts of a pField object}
\source{
Function adapted from "proxytype.R" in paleolibary/src/.
}
\usage{
\method{[}{pField}(x, p1, p2, ...)
}
\arguments{
\item{x}{a \code{"pField"} object.}

\item{p1}{possible row indices for the subset: \code{x[p1, ]}.}

\item{p2}{possible column indices for the subset: \code{x[, p2]}.}

\item{...}{further arguments passed on to the generic method \code{"["};
currently not needed.}
}
\value{
the original object with the replaced subset \code{x[p1, p2]}, or
the subset \code{x[p1, p2]}. For the latter, depending on the provided
indices, this can be an object of class \code{"pField"}, an object of class
\code{"pTs"}, or simply a numeric vector; see details.
}
\description{
Method for extracting or replacing a subset of a \code{"pField"} object.
}
\details{
For subsetting, the class of the returned object depends on the specified
indices:
\describe{
\item{\code{x[p1, ]}}{a \code{"pField"} object.}
\item{\code{x[, p2]}}{a \code{"pTs"} object if \code{p2} is of length 1
(single site) or if the subset is not compatible with a \code{"pField"}
object (i.e. the subset sites cover more than one latitude, but, for each,
do not cover the full range of longitudes).}
\item{\code{x[p1, p2]}}{same as before for the time subset \code{p1}.}
\item{\code{x[p1]}}{a numeric vector of length \code{length(p1)} where the
time and lat-lon information are lost with a warning.}
\item{\code{x[]} or \code{x[,]}}{the original object \code{x}.}
}
}
\seealso{
\code{?[.pTs}
}
\author{
Thomas Laepple; adapted by Thomas Münch
}
