% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing-methods.R
\name{[.pTs}
\alias{[.pTs}
\title{Extract or replace parts of a pTs object}
\source{
Function adapted from "proxytype.R" in paleolibary/src/.
}
\usage{
\method{[}{pTs}(x, p1, p2, ...)
}
\arguments{
\item{x}{a \code{"pTs"} object.}

\item{p1}{possible row indices for the subset: \code{x[p1, ]}.}

\item{p2}{possible column indices for the subset: \code{x[, p2]}.}

\item{...}{further arguments passed on to the generic method \code{"["};
currently not needed.}
}
\value{
the original object with the replaced subset \code{x[p1, p2]}, or
the subset \code{x[p1, p2]}. For the latter, the return value is an object
of class \code{"pTs"}, unless subset is \code{x[p1]} in which case it is a
numeric vector of length \code{length(p1)} where the time and lat-lon
information are lost with a warning.
}
\description{
Method for extracting or replacing a subset of a \code{"pTs"} object.
}
\seealso{
\code{?[.pField}
}
\author{
Thomas Laepple; adapted by Thomas Münch
}
