% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{summary.pField}
\alias{summary.pField}
\title{pField summary}
\usage{
\method{summary}{pField}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"pField"}.}

\item{...}{additional arguments; currently not implemented.}
}
\description{
Method function to produce the summary for a \code{"pField"} object.
}
\author{
Thomas Münch
}
