% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{summary.pTs}
\alias{summary.pTs}
\title{pTs summary}
\usage{
\method{summary}{pTs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"pTs"}.}

\item{...}{additional arguments; currently not implemented.}
}
\description{
Method function to produce the summary for a \code{"pTs"} object.
}
\author{
Thomas Münch
}
