% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-fields.R
\name{ApplySpace}
\alias{ApplySpace}
\alias{applyspace}
\title{Apply a function across space}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
ApplySpace(data, FUN, ...)
}
\arguments{
\item{data}{a \code{"pField"} or \code{"pTs"} object.}

\item{FUN}{the function to be applied.}

\item{...}{further arguments passed on to \code{FUN}.}
}
\value{
a single \code{"pTs"} time series with the results of the function
  applied.
}
\description{
Apply a function across the spatial dimension of a \code{"pField"} or
\code{"pTs"} object by using \code{apply} on the rows of the object. The
result of this gives a \code{"pTs"} vector object, i.e. a single time
series.
}
\details{
Input columns which only contain NA values are stripped befor applying the
function. If only one non-NA column is present, this column is returned
unchanged with a warning.
}
\examples{
x <- pField(data = array(rnorm(10 * 10 * 100), dim = c(10, 10, 100)),
            time = 1 : 100, lat = 1 : 10, lon = 1 : 10)
ApplySpace(x, sd)

x <- x[, 1 : 15]
ApplySpace(x, sd)
}
\author{
Thomas Laepple
}
