% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-fields.R
\name{ApplyTime}
\alias{ApplyTime}
\alias{applytime}
\title{Apply a function across time}
\source{
Function copied from "basis.R" in paleolibary/src/.
}
\usage{
ApplyTime(data, FUN, newtime = NULL, ...)
}
\arguments{
\item{data}{a \code{"pField"} object.}

\item{FUN}{the function to be applied.}

\item{newtime}{Specify the observation time point corresponding to the
resulting \code{"pField"} object. For \code{NULL} (the default), the average
of the time points in \code{data} is used.}

\item{...}{further arguments passed on to \code{FUN}.}
}
\value{
a \code{"pField"} object with the results of the function applied.
}
\description{
Apply a function across the temporal dimension of a \code{"pField"} object by
using \code{apply} on the columns of the object. The result of this gives
a \code{"pField"} object with only one time step.
}
\examples{
x <- pField(data = array(rnorm(10 * 10 * 100), dim = c(10, 10, 100)),
            time = 1 : 100, lat = 1 : 10, lon = 1 : 10)
ApplyTime(x, mean)
}
\author{
Thomas Laepple
}
