% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{cor.pTs}
\alias{cor.pTs}
\title{Correlation of point with field}
\source{
Function based on "mat.pField.R" in paleolibary/src/.
}
\usage{
cor.pTs(point, field, debug = FALSE, ...)
}
\arguments{
\item{point}{a \code{"pTs"} object.}

\item{field}{a \code{"pField"} object.}

\item{debug}{\code{point} and \code{field} are brought onto the same time
basis. If \code{debug = TRUE}, this new time basis is printed as a message.}

\item{...}{further arguments passed on to the base R correlation function
\code{\link[stats]{cor}}, such as the \code{use} argument.}
}
\value{
the correlation(s) between \code{point} and \code{field}. The class
of the result depends on the class of \code{field}, so a \code{"pField"}
object usually.
}
\description{
Calculate the correlation between an object of class \code{"pTs"} and an
object of class \code{"pField"}.
}
\examples{
point <- rnorm(100)
field <- pField(rnorm(100 * 10 * 10), time = 1 : 100,
                lat = 1 : 10, lon = 1 : 10)

correlation <- cor.pTs(point, field)
}
\author{
Thomas Laepple, modified by Thomas Münch
}
