% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{is.pField}
\alias{is.pField}
\title{pField object}
\usage{
is.pField(object)
}
\arguments{
\item{object}{object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{object} is a
pField or not.
}
\description{
Test for pField object
}
\author{
Thomas Laepple
}
