% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{is.pTs}
\alias{is.pTs}
\title{pTs object}
\usage{
is.pTs(object)
}
\arguments{
\item{object}{object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{object} is a
\code{"pTs"} object or not.
}
\description{
Check if object is a \code{"pTs"} object.
}
\examples{
# Some pTs object
x <- pTs(lat = -75, lon = 0, time = 1 : 10)

is.pTs(x)
is.pTs(1 : 10)
}
\author{
Thomas Laepple
}
