% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-methods.R
\name{pField2df}
\alias{pField2df}
\title{Convert field to data frame}
\usage{
pField2df(data, lat.min = NULL, lat.max = NULL, lon.min = NULL, lon.max = NULL)
}
\arguments{
\item{data}{a \code{"pField"} or \code{"pTs"} object with one timestep.}

\item{lat.min}{set the minimum latitude for the output; per default the
minimum in \code{data} is used.}

\item{lat.max}{set the maximum latitude for the output; per default the
maximum in \code{data} is used.}

\item{lon.min}{set the minimum longitude for the output; per default the
minimum in \code{data} is used.}

\item{lon.max}{set the maximum longitude for the output; per default the
maximum in \code{data} is used.}
}
\value{
A data frame with the three columns \code{lat} (latitudes),
  \code{lon} (longitudes) and \code{dat} (the values of \code{data} at the
  coordinate positions).
}
\description{
Convert a static \code{"pField"} or \code{"pTs"} object, i.e. including only
one time step, to a data frame with data, latitude and longitude columns,
with the possibility to cut out a specified latitude-longitude region.
}
\examples{
# Create a pfield object covering two latitudes, three longitudes and
# four time steps, let data values increase with latitude and time only:
lat <- seq(-75, -80, -5)
lon <- c(0, 135, 215)
time <- 1 : 4
space <- c(1, 1, 1, 2, 2, 2)
spacetime <- c(space, 10 * space, 100 * space, 1000 * space)

x <- pField(data = spacetime, lat = lat, lon = lon, time = time)

# Average across time
x.avg <- ApplyTime(x, mean)

# Convert to data frame
x.df <- pField2df(x.avg)
x.df

# Cut out some region
x.df <- pField2df(x.avg, lat.min = -75, lon.min = 0, lon.max = 200)
x.df

# Subset incompletely to create a pTs object
x.pts <- x[, 1 : 4]
x.avg <- ApplyTime(x.pts, mean)

# Convert to data frame
x.df <- pField2df(x.avg)
x.df # the same as x.df[1 : 4,] from above
}
\author{
Thomas Münch
}
